Separated control layer from mover script
