+++ Assets/Game/Characters/Enemy.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: f5102e87e7b2e72669bb65244b7c3d2e
-folderAsset: yes
-DefaultImporter:
+guid: d32002d2cb0f24e41a45158d7bc9dacf
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
