+++ Assets/Game/Characters/Player.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 1136244085944684601}
   - component: {fileID: 122433474194075215}
   - component: {fileID: 442536700}
+  - component: {fileID: 409057543690640014}
   - component: {fileID: 442536701}
   - component: {fileID: 4945303585339192480}
   m_Layer: 0
@@ -87,6 +88,18 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   target: {fileID: 0}
+--- !u!114 &409057543690640014
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9075820937765519782}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2e3f9c178fba7924d9ac4730314a6da4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!195 &442536701
 NavMeshAgent:
   m_ObjectHideFlags: 0
