+++ Assets/Scenes/Sandbox.unity
@@ -2368,3 +2368,72 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: f63fb9b5a1dbe46c297553be9728c32b, type: 3}
+--- !u!1001 &5450128774595493570
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 599378995386237205, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 4.85
+      objectReference: {fileID: 0}
+    - target: {fileID: 599378995386237205, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.92
+      objectReference: {fileID: 0}
+    - target: {fileID: 599378995386237205, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 3.87
+      objectReference: {fileID: 0}
+    - target: {fileID: 599378995386237205, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 599378995386237205, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 599378995386237205, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 599378995386237205, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 599378995386237205, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 5
+      objectReference: {fileID: 0}
+    - target: {fileID: 599378995386237205, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 599378995386237205, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 599378995386237205, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6203470901424545297, guid: d32002d2cb0f24e41a45158d7bc9dacf,
+        type: 3}
+      propertyPath: m_Name
+      value: Enemy
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: d32002d2cb0f24e41a45158d7bc9dacf, type: 3}
