+++ Assets/Scripts/Control.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d0f2a61a319684b26a3bc62d2c8ad85e
+guid: 05bd4f132543b154090d390d58979f06
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
