+++ Assets/Scripts/Movement/Mover.cs
@@ -5,44 +5,33 @@ using UnityEngine.AI;
 
 public class Mover : MonoBehaviour
 {
-    [SerializeField] Transform target;
-    NavMeshAgent agent;
+  [SerializeField] Transform target;
+  NavMeshAgent agent;
 
-    // Start is called before the first frame update
-    void Start()
-    {
-        agent = gameObject.GetComponent<NavMeshAgent>();
-    }
+  // Start is called before the first frame update
+  void Start()
+  {
+    agent = gameObject.GetComponent<NavMeshAgent>();
+  }
 
-    // Update is called once per frame
-    void Update()
-    {
-        if(Input.GetMouseButton(0))
-        {
-            MoveToCursor();
-        }
-        UpdateAnimator();
-        //Debug.DrawRay(lastRay.origin, lastRay.direction * 100);
+  // Update is called once per frame
+  void Update()
+  {
+    UpdateAnimator();
+    //Debug.DrawRay(lastRay.origin, lastRay.direction * 100);
+  }
 
-    }
+  public void MoveTo(Vector3 destination)
+  {
+  agent.SetDestination(destination);
+  }
 
-    private void MoveToCursor()
-    {
-        Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
-        RaycastHit hit;
-        bool hasHit = Physics.Raycast(ray, out hit);
-        if(hasHit)
-        {
-            agent.SetDestination(hit.point);
-        }
-    }
-    private void UpdateAnimator()
-    {
-        Vector3 velocity = GetComponent<NavMeshAgent>().velocity;
-        Vector3 localVelocity = transform.InverseTransformDirection(velocity);
-        float speed = localVelocity.z;
+  private void UpdateAnimator()
+  {
+    Vector3 velocity = GetComponent<NavMeshAgent>().velocity;
+    Vector3 localVelocity = transform.InverseTransformDirection(velocity);
+    float speed = localVelocity.z;
 
-        GetComponent<Animator>().SetFloat("forwardSpeed", speed); //string reference
-    }
-
+    GetComponent<Animator>().SetFloat("forwardSpeed", speed); //string reference
+  }
 }
