--- Assets/UI/Inventory.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d0e81ed0c75b4097c844c9f6bde881dc
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
